moved stats calculation to separate functions
