+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./cover";
 export * from "./errors";
 export * from "./htttp";
+export * from "./stats";
 export * from "./utils";
